+++ FogApi/Public/Get-FogGroups.ps1
@@ -1,0 +1,32 @@
+function Get-FogGroups {
+    <#
+    .SYNOPSIS
+    Returns all fog groups
+    
+    .DESCRIPTION
+    Gets all the fog groups and returns them in an object
+    
+    .EXAMPLE
+    $groups = Get-FogGroups
+    
+    .NOTES
+    A group object does not contain membership information, you need to filter groupassociations to find membership
+    but this will give you the id to search for
+    #>
+    [CmdletBinding()]
+    param (
+        
+    )
+    
+    begin {
+        
+    }
+    
+    process {
+        Parameter description
+    }
+    
+    end {
+        
+    }
+}